Moving estimate logic to own container component
